Prompts/requirements.txt
langchain
langchain-core
langchain-openai
python-dotenv
streamlit
#pip install -r requirements.txt 